Re-organise authentication related components into one module. Deliberately keeping other fixes minimal for now.
