+++ src/commons/HOCs/withAuthentication.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from "react";
 import { connect } from "react-redux";
 import { Redirect } from "react-router-dom";
 
-import { registerAuthStateObserver } from "actions/auth";
+import { registerAuthStateObserver } from "modules/auth/actions";
 import Progress from "components/Progress";
 
 export default function withAuthentication(Comp, FallbackComp) {
