--- src/containers/pages/ResetPasswordPage.js
@@ -1,7 +1,0 @@
-import { connect } from "react-redux";
-import { withRouter } from "react-router-dom";
-
-import ResetPasswordPage from "pages/ResetPasswordPage";
-
-const mapStateToProps = ({ user }) => ({ user });
-export default connect(mapStateToProps)(withRouter(ResetPasswordPage));
