+++ src/modules/auth/actions.js
@@ -1,8 +1,7 @@
 import { handleError } from "utils";
 import { asyncAction, editAction } from "commons/utils/action-creators";
+import { fetchProfile } from "actions/profile";
 import firebase from "firebase.js";
-
-import { fetchProfile } from "./profile";
 
 export const USER_SIGN_IN = asyncAction("USER_SIGN_IN");
 export const USER_NOT_SIGN_IN = "USER_NOT_SIGN_IN";
