+++ src/modules/auth/components/AppDrawerAuth.js
@@ -5,7 +5,7 @@ import List from "@material-ui/core/List";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemText from "@material-ui/core/ListItemText";
 
-import Progress from "./Progress.js";
+import Progress from "components/Progress";
 
 const AppDrawerAuth = ({
   notAuthenticated,
