+++ src/modules/auth/containers/ResetPasswordForm.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
-import { resetPassword } from "actions/auth";
-import ResetPasswordForm from "components/ResetPasswordForm";
+import { resetPassword } from "../actions";
+import ResetPasswordForm from "../components/ResetPasswordForm";
 
 const mapStateToProps = ({ user: { loading } }) => ({ loading });
 const mapDispatchToProps = { resetPassword };
