+++ src/modules/auth/containers/ResetPasswordPage.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 
-import SignInPage from "pages/SignInPage";
+import ResetPasswordPage from "../pages/ResetPasswordPage";
 
 const mapStateToProps = ({ user }) => ({ user });
-export default connect(mapStateToProps)(withRouter(SignInPage));
+export default connect(mapStateToProps)(withRouter(ResetPasswordPage));
