+++ src/modules/auth/containers/SignInPage.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 
-import PasswordChangePage from "pages/PasswordChangePage";
+import SignInPage from "../pages/SignInPage";
 
 const mapStateToProps = ({ user }) => ({ user });
-export default connect(mapStateToProps)(withRouter(PasswordChangePage));
+export default connect(mapStateToProps)(withRouter(SignInPage));
