+++ src/modules/auth/index.js
@@ -1,0 +1,4 @@
+export { default as SignInPage } from "./containers/SignInPage";
+export { default as ResetPasswordPage } from "./containers/ResetPasswordPage";
+export { default as PasswordChangePage } from "./pages/PasswordChangePage";
+export { default as AppDrawerAuth } from "./containers/AppDrawerAuth";
