+++ src/modules/auth/pages/PasswordChangePage.js
@@ -2,7 +2,8 @@ import React from "react";
 
 import AppFrame from "components/AppFrame";
 import Page from "components/Page";
-import PasswordChangeForm from "containers/components/PasswordChangeForm";
+
+import PasswordChangeForm from "../components/PasswordChangeForm";
 
 const PasswordChangePage = () => (
   <AppFrame title="Change Your Password">
