+++ src/modules/auth/pages/ResetPasswordPage.js
@@ -3,8 +3,9 @@ import { Link, Redirect } from "react-router-dom";
 import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
 
-import ResetPasswordForm from "containers/components/ResetPasswordForm";
 import Page from "components/Page";
+
+import ResetPasswordForm from "../components/ResetPasswordForm";
 
 const ResetPasswordPage = ({ history, user }) =>
   !user.hasTriggeredReset ? (
