+++ src/modules/auth/pages/SignInPage.js
@@ -3,8 +3,9 @@ import { Link, Redirect } from "react-router-dom";
 import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
 
-import SignInForm from "containers/components/SignInForm";
 import Page from "components/Page";
+
+import SignInForm from "../components/SignInForm";
 
 const SignInPage = ({ history, user }) =>
   !user.data ? (
